requirements.txt
requests
beautifulsoup4
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
huggingface_hub